requirements.txt
streamlit
pdfminer.six
python-docx
pytesseract
Pillow
spacy
keybert
bertopic
scikit-learn
wordcloud
matplotlib
numpy
plotly
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
